requirements.txt
flask
flask-cors
flask-sqlalchemy
psycopg2-binary
flask-jwt-extended
pytest
coverage
sqlalchemy-utils
python-dotenv
build
